some minor changes and rest from akanksha's branch

+++ src/main/java/com/mealon/MEAL_ON/service/MessServiceImpl.java
@@ -10,6 +10,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.mealon.MEAL_ON.dao.MessDAO;
 import com.mealon.MEAL_ON.model.Mess;
+import com.mealon.MEAL_ON.model.Student;
 
 @Service
 public class MessServiceImpl implements MessService{
@@ -18,18 +19,47 @@ public class MessServiceImpl implements MessService{
 	
 	@Transactional
 	@Override
-	public String add(String name, String password, String messadmin) {
-		String result = null;
-		Mess newMess = toMess(name, password, messadmin, 0);
+	public Integer add(String name, String password, String messadmin) {
+		/* Fails if same Mess name already exists in database
+		 * Fails also when DAO is unable to save::: not to be informed to the customer
+		 * Success if Mess name is unique
+		 */
+		Integer result = 0;
+		Mess mess = messDAO.findByName(name);
+		if(mess == null) {
+			Mess newMess = toMess(name, password, messadmin, 0);
+			try {
+				messDAO.save(newMess);
+				result = newMess.getMessid();
+			}
+			catch (Exception e) {
+				e.printStackTrace();
+			}
+		}
+		return result;
+	}
+	
+	@Transactional
+	@Override
+	public Boolean check(Integer mess_id , String password) {
+		/* Fails if password of mess does not match with the saved password
+		 * Success if password matches
+		 */
+		Boolean result = false;
 		try {
-			messDAO.save(newMess);
-			result = "Added successfully with mess_id : " + newMess.getMessid();
+			Mess mess = messDAO.findByMessid(mess_id);
+			String pass = mess.getPassword();
+			System.out.println(mess.getName());
+			if(pass.equals(password)) {
+				result = true;
+			}
 		}
 		catch (Exception e) {
 			e.printStackTrace();
 		}
 		return result;
 	}
+	
 	@Transactional
 	@Override
 	public List<Mess> get() {
@@ -46,14 +76,10 @@ public class MessServiceImpl implements MessService{
 	
 	@Transactional
 	@Override
-	public Mess get(String name, String password) {
+	public Mess get(Integer mess_id) {
 		Mess mess = null;
 		try {
-			mess = messDAO.findByName(name);
-			String pass = mess.getPassword();
-			if(!pass.equals(password)) {
-				mess = null;
-			}
+			mess = messDAO.findByMessid(mess_id);
 		}
 		catch (Exception e) {
 			e.printStackTrace();
@@ -62,35 +88,49 @@ public class MessServiceImpl implements MessService{
 	}
 	@Transactional
 	@Override
-	public String update(Integer mess_id, String name, String password, String messadmin, Integer rate) {
-		Mess newMess = toMess(name, password, messadmin, rate);
-		String result = null;
-		try {
-			messDAO.save(newMess);
-			result = "Successfully updated";
-		}
-		catch (Exception e) {
-			e.printStackTrace();
+	public Boolean update(Integer mess_id, String name, String password, String messadmin, Integer rate) {
+		/* Fails if Mess with given name does not exists in database
+		 * Success if Mess is updated
+		 */
+		Boolean result = false;
+		Mess mess = get(mess_id);
+		if(mess != null) {
+			Mess newMess = toMess(name, password, messadmin, rate);
+			newMess.setMessid(mess_id);
+			try {
+				messDAO.save(newMess);
+				result = true;
+			}
+			catch (Exception e) {
+				e.printStackTrace();
+			}
 		}
 		return result;
 	}
 	
 	@Transactional
 	@Override
-	public String delete(int mess_id, String name, String password) {
+	public Boolean delete(int mess_id, String name, String password) {
+		/* Fails if password does not match,
+		 * Fails if name does not match.
+		 * Success if mess with the name does not exists,
+		 * Success if mess is deleted.
+		 */
+		Boolean result = false;
 		Mess mess = null;
-		String result = "Does not exists";
 		try {
 			Optional<Mess> optional = messDAO.findById(mess_id);
-			mess = (Mess) optionalToList(optional);
-			String Password = mess.getPassword();
-			String Name = mess.getName();
-			if(Password.equals(password) && Name.equals(name)) {
-				messDAO.deleteById(mess_id);
-				result = "Deleted successfuly";
+			if(optional.isPresent()) {
+				mess = (Mess) optionalToList(optional);
+				String Password = mess.getPassword();
+				String Name = mess.getName();
+				if(Password.equals(password) && Name.equals(name)) {
+					messDAO.deleteById(mess_id);
+					result = true;
+				}
 			}
 			else {
-				result = "Invalid name or password";
+				result = true;
 			}
 		}
 		catch (Exception e) {
@@ -101,6 +141,7 @@ public class MessServiceImpl implements MessService{
 	
 	private Mess toMess(String name, String password, String messadmin, Integer rate) {
 		Mess newMess = new Mess();
+		//mess_id will be auto incremented
 		//newMess.setMessid(mess_id);
 		newMess.setName(name);
 		newMess.setPassword(password);
